sys: 64 bit compatibility
Changes other than DEBUG/printf formatting or different types for void pointer casting
* vfs.h: Increased default vfs buffer sizes for 64 bit
* bloom.h: Fixed typedef for the hashfp_t function pointer
* bytes.h: Check if socklen_t is already defined
* priority_queue_t: Replaced `-1U` literal with PRIORITY_QUEUE_DATA_SIGNALING define

+++ sys/suit/handlers_command_seq.c
@@ -330,7 +330,7 @@ static inline void _print_download_progress(suit_manifest_t *manifest,
     (void)manifest;
     (void)offset;
     (void)len;
-    DEBUG("_suit_flashwrite(): writing %u bytes at pos %u\n", len, offset);
+    DEBUG("_suit_flashwrite(): writing %u bytes at pos %u\n", (unsigned)len, (unsigned)offset);
 #if defined(MODULE_PROGRESS_BAR)
     if (image_size != 0) {
         char _suffix[7] = { 0 };
